refactor(ConnectionTest): move to lib
